requirements.txt
sentence-transformers
faiss-cpu
pypdf
openai
python-dotenv
